docs: improving docs and examples
